+++ src/runtime/scope/index.ts
@@ -1,0 +1,2 @@
+export * from './test_case_scope'
+export * from './test_run_scope'