+++ src/support_code_library_builder/context.ts
@@ -1,0 +1,3 @@
+export interface IContext<ParametersType = any> {
+  readonly parameters: ParametersType
+}